lang_chain_email_agent/requirements.txt
# requirements.txt

# Google API client libraries
google-api-python-client>=2.70.0
google-auth>=2.6.6
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6

# LangChain and OpenAI
langchain>=0.0.100
openai>=0.27.0

# Additional dependencies
httplib2>=0.21.0
six>=1.16.0
uritemplate>=4.1.1


